Implemented helper for assigning settings the HydraRingCalculationInput, which now contains settings properties for design tables, numerics and hydraulic models. Related to Issue [WTI-919]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@6425 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 242033e72f79b68d1d2a3a4cab271dcabc178ff1
Former-commit-id: 1bb23f4e2da46c582bf4bbb174ee2e657c28c395

+++ Ringtoets/Common/src/Ringtoets.Common.IO/HydraRing/NumericsSettingsProvider.cs
@@ -61,13 +61,13 @@ namespace Ringtoets.Common.IO.HydraRing
         /// <param name="failureMechanismType">The <see cref="HydraRingFailureMechanismType"/> to obtain the <see cref="NumericsSetting"/> for.</param>
         /// <returns>A new <see cref="Dictionary{T, T}"/> where the key is the submechanism id, and the value is
         /// the <see cref="NumericsSetting"/> containing values corresponding to the provided failure mechanism type and location id.</returns>
-        public Dictionary<long, NumericsSetting> GetNumericsSettings(long locationId, HydraRingFailureMechanismType failureMechanismType)
+        public Dictionary<int, NumericsSetting> GetNumericsSettings(long locationId, HydraRingFailureMechanismType failureMechanismType)
         {
             var failureMechanismDefaults = new FailureMechanismDefaultsProvider().GetFailureMechanismDefaults(failureMechanismType);
             var subMechanismIds = failureMechanismDefaults.SubMechanismIds;
             var mechanismId = failureMechanismDefaults.MechanismId;
 
-            var numericsSettings = new Dictionary<long, NumericsSetting>();
+            var numericsSettings = new Dictionary<int, NumericsSetting>();
 
             foreach (var subMechanismId in subMechanismIds)
             {
